+++ protocols/refresh/fake.go
@@ -1,4 +1,4 @@
-package session
+package refresh
 
 import (
 	"crypto/rand"
