+++ protocols/refresh/result.go
@@ -2,10 +2,9 @@ package refresh
 
 import (
 	"github.com/taurusgroup/cmp-ecdsa/pkg/party"
-	"github.com/taurusgroup/cmp-ecdsa/pkg/session"
 )
 
 type Result struct {
-	Session *session.Session
+	Session *Session
 	Secret  *party.Secret
 }
